Backport _PyImport_SwapPackageContext unconditionally.
I am proposing my change to _memimporter that made it compile with a backported _PyImport_SwapPackageContext that is manually provided since it is not exported from pythonXY.dll itself. Will need to manually test the copy on this repo as well first to see which one works better in 3.12 for my needs.